Move dejs.ts to mod.ts
