+++ README.md
@@ -21,7 +21,7 @@
 ## Usage
 
 ```ts
-import * as dejs from 'https://deno.land/x/dejs@0.3.0/mod.ts';
+import * as dejs from 'https://deno.land/x/dejs@0.3.1/mod.ts';
 ```
 
 - `renderFile(filePath: string, params: Params): Promise<Reader>`
